+++ changelog.md
@@ -15,6 +15,7 @@
 - plot.period now displays day plots (zoom = 'day') for days in the past - starting next day at 00:00 after tmin ('24h' = today, '48h' = yesterday ...). 
 - new widget status.customstyle applies custom CSS styles to any linked widget or html element if values of a trigger item match the given conditions.
 - refactored weather app and widgets weather.map and weather.mapslides after change of Tagesschau content management. (Thanks to Tagesschau technical management for hints on new API).
+- new "style" option for basic slider allows applying CSS styles to the slider track
 
 ### Other New Features
 - weather service pirateweather.net as replacement for darksky.net (thanks to aschwith)